pyproject.toml
[tool.poetry]
name = "coursework-1"
version = "0.1.0"
description = ""
authors = ["Anna <azeleranskaya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openpyxl = "^3.1.5"
pandas = "^2.2.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.lint.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
mypy = "^1.13.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '.git'

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'